added 2fa support

+++ models/file.model.go
@@ -15,7 +15,7 @@ type File struct {
 	Size      int64     `gorm:"type:bigint"`
 	Starred   *bool     `gorm:"default:false"`
 	Depth     *int      `gorm:"type:integer"`
-	UserID    int       `gorm:"type:bigint;not null"`
+	UserID    int64     `gorm:"type:bigint;not null"`
 	Status    string    `gorm:"type:text"`
 	ParentID  string    `gorm:"type:text;index"`
 	Parts     *Parts    `gorm:"type:jsonb"`
